fix: extract session manifest helpers + ClickHouse env guard
Extracted `_atomic_index_update`, `_finalize_session`, and
`_re_session_id_pattern` from `cli/commands/sweep.py` to a new
`cli/_session_manifest.py` module. Reduces sweep.py to 1889 lines,
ratio 17.49 (under MAX_RATIO=18). The bloat-gate mutation test
fixture (which creates a ratio-20 file) now fires correctly again.

ClickHouse env guard: `_run` now skips the trades parquet prefetch
when `CLICKHOUSE_PASSWORD` is unset, writing the reason to stderr
and passing `trades_parquet_path=None` to subprocesses. CI unit
tests (`test_sweep_parity_guard`) mock `run_window_in_subprocess`
so they don't actually need ClickHouse; the env-missing branch
lets them reach the parity-check assertions. Real CLI usage always
has the env set.

Issue 62 Surfaces updated to include the new
`backtest_simulator/cli/_session_manifest.py` path.

Module budgets: sweep.py 1545 -> 1450, _session_manifest.py 200.
File-size MAX_RATIO 20 -> 18 (mutation test now passes).
