+++ .github/module_budgets.json
@@ -58,11 +58,12 @@
   "backtest_simulator/cli/_pipeline.py": 760,
   "backtest_simulator/cli/_metrics.py": 245,
   "backtest_simulator/cli/_run_window.py": 760,
+  "backtest_simulator/cli/_session_manifest.py": 200,
   "backtest_simulator/cli/commands/__init__.py": 10,
   "backtest_simulator/cli/commands/run.py": 420,
   "backtest_simulator/cli/_signals_builder.py": 500,
   "backtest_simulator/cli/_stats.py": 440,
-  "backtest_simulator/cli/commands/sweep.py": 1545,
+  "backtest_simulator/cli/commands/sweep.py": 1450,
   "backtest_simulator/cli/commands/enrich.py": 60,
   "backtest_simulator/cli/commands/test.py": 80,
   "backtest_simulator/cli/commands/lint.py": 60,
